scripts/requirements.txt
pandas
psycopg2-binary
sqlalchemy
PyPDF2
python-docx
textblob
wordcloud
